Fixed date and time formatting.  Closes #99
The reason I did not add the periods between am and pm is because there are no native ways to do that with any of the existing datetime libraries I've seen.  We use javascript moment.js and they don't provide the periods in their extensive list of formatting options.  Ruby does not support the periods either.  I say we use what the rest of the web uses and keep the periods out.