Live Previewing Redone
Live previewing used to crash the browser.  We now save a another slide in the database with a corresponding negative id.  Then, we just update that thing, and reload the iframe that is previewing it.  It is a little tricky, and there are still some bugs when you are adding and removing images, but it is much better that what we have now, and those bugs can be for another time.

Sincerely,
James